refactor js Todo List
