+++ src/render.js
@@ -1,17 +1,14 @@
+import { listContainer, createList } from './createTask.js';
 // eslint-disable-next-line import/no-cycle
-import { deleteOne, createList, listContainer } from './index.js';
-// eslint-disable-next-line import/no-cycle
-import addCheckListener from './status.js';
+import { addCheckListener, deleteOne } from './index.js';
 
-// render each task to follow the index.
 const renderTasks = (taks) => {
   listContainer.innerHTML = '';
   taks.forEach((t, i) => {
     t.index = i + 1;
     const taskNode = createList(t);
-    addCheckListener(taskNode, i);
     deleteOne(taskNode, i);
+    addCheckListener(taskNode, i);
   });
 };
-
 export default renderTasks;
